Address review feedback from guykogus
- Move deprecated functions to Deprecated/FileManagerDeprecated.swift
- Make test functions throwing instead of using do-catch
- Move #if conditions outside function definitions for withFilename test
- Update CHANGELOG to list deprecated methods
